Rename Modal to BranchModal
